refactor unread_count polling so it doesn't slam a backed-up server
fixes: CNVS-21942

let's say you had a server that was backed up and
taking a really long time to respond. as is, this
code would bombard it with a bunch of requests
to poll the unread_count.

this changes it so that rather than sending a request
every 30 seconds, it sends a request 30 seconds
*after the last response came back* so it is nice
to a server that is overloaded.

test plan:

* load the new UI and look at the unread count
  in the global nav.
* in another tab/browser, send yourself a new inbox
  message
* within 30 seconds, the original tab you had open
  should indicate that you have a new unread message

Change-Id: I41497cb87e57b997dfb71eb5604ca13f5c6094b6
Reviewed-on: https://gerrit.instructure.com/58477
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>