Appium Specs for Mobile Apps
This includes a basic set of common methods and environment variables
to be used for both iOS and Android platforms with Appium. Includes the
inclusion of the appium_lib gem in Gemfile.d and a minor modification
to spec/spec_helper.rb to facilitate control of seeding the ID for
DeveloperKeys for Mobile Verify.

To Test:
  - remove skip statements in examples
  - uncomment the call to appium_init()
  - verify appium_lib gem installs during bundle update
  - spin up an Appium server
  - run spec with the QA Nexus 7 connected

Change-Id: Ia9830aeaa0b91635f5739958466e3176f25bf2bb
Reviewed-on: https://gerrit.instructure.com/58562
Tested-by: Jenkins
Reviewed-by: Heath Hales <hhales@instructure.com>
Reviewed-by: Ben Bolton <bbolton@instructure.com>
QA-Review: Ben Bolton <bbolton@instructure.com>
Product-Review: Ben Bolton <bbolton@instructure.com>