added anonymous grading feature flag
fixes CNVS-21535

For now this can just be a plain feature flag because feature
flags implement the "cascading state" functionality we want for
this setting. We can change the UI on this or convert it to a
new type of setting later if we think that's important enough.

test plan:
 - Anonymous grading feature flag exists as a feature currently
 under development (only visible to site admins until it is
 turned on). Nothing else should be affected by this
 flag's setting yet.

Change-Id: I5136265cd46af47ea48482b2ab5015112e079557
Reviewed-on: https://gerrit.instructure.com/60871
Tested-by: Jenkins
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Dan Minkevitch <dan@instructure.com>