[NUI][a11y] Makes global nav trays populate when opening
This makes it so keyboard only and screenreader users can see
the list of courses, accounts, and groups in the new UI. It does
so by checking to see if the items for the tray have been loaded
or are loading.  If not it will trigger them to load when the
tray itself opens.

fixes CNVS-22736

Test Plan:
  - Using only the keyboard open the Course, Admin, and Group
    global menu items.
  - The list of courses, accounts, and groups should be populated
    or triggered to populate (and finish loading)
  - Do the same with screenreaders

Change-Id: I974c32023de01dbca43c36a51e23e6308931291e
Reviewed-on: https://gerrit.instructure.com/61908
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Tested-by: Jenkins
Product-Review: Clay Diffrient <cdiffrient@instructure.com>