implement switch to points/percentage toggle
clicking the 'Switch to Points/Percentages' toggle
in the 'Total' column will change the display of
total grades. a warning dialog will show up when
clicking the toggle if the user has not previously
checked the 'dont show for this course again' option.
if the use has checked that option before, the dialog
will not show up.

closes CNVS-22037

test plan:

1. verify the warning dialog shows up when clicking the
toggle. when the dialog shows up, click 'don't show for
this course again' and continue. any subsequent clicks
on the toggle should not result in a warning dialog showing.
2. on the assignments page, click the settings cog and turn
on 'Weight final grade based on assignment groups'. when
this setting is on, the 'Switch to Points/Percentages' toggle
should not be displayed in the gradebook.
3. Verify that the total shows as '-' if there are no
submissions, regardless of display being Points or Percentage.
4. Verify that the gradebook remembers preferences accross
page refreshes for a) display as points/percentage, and b)
whether or not the warning dialog should show up (you will need
to delete the '_id_course_id_warned_about_totals_display' key
from localStorage if you want the warning dialog to show up
after you've already hidden it).

Change-Id: I60a557ca952d6f5025e2fb1c6fda94bf1d5493ed
Reviewed-on: https://gerrit.instructure.com/63178
Tested-by: Jenkins
Reviewed-by: Dylan Ross <dross@instructure.com>
QA-Review: Jason Carter <jcarter@instructure.com>
Product-Review: Spencer Olson <solson@instructure.com>