fix css & js on 404.html page
fixes: CNVS-22835

this is just a quick fix so it works like it used
to.  We really should re-do that entire page so
it does not fragilely depend on css and js from
the rails app & load it in a way that is different
from everything else.  We'll do that as a separate
ticket.

Change-Id: I362b66f968015520d9b44e787d84021fa258218b
Reviewed-on: https://gerrit.instructure.com/62092
Reviewed-by: Rob Orton <rob@instructure.com>
Tested-by: Jenkins
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>