support for accessibility for expand and collapsing discussions
Previously there was a header tag and a link button that the user could
click on to collapse but the readers weren't picking it up correctly.

This commit removes the collapsibility from the header and just assigns it
to the link button which can then be focused on for the screen readers.

Also, removed is the mark as read tooltip since it breaks accessibility.

fixes CNVS-22058

test plan:
- Go to the discussion_topics show page: /courses/1/discussion_topics/1
  - If the discussion has sub discussions then use the keyboard to navigate
    to the next element past the Mark as Read/Unread button and ensure that the
    VO, NVDA, and JAWS messages make sense.

Change-Id: Id26a2d63f77b9b8ca38b645207c63cae40887aa3
Reviewed-on: https://gerrit.instructure.com/64679
Reviewed-by: Matt Berns <mberns@instructure.com>
Tested-by: Jenkins
Product-Review: Aaron Cannon <acannon@instructure.com>
QA-Review: Michael Hargiss <mhargiss@instructure.com>