[a11y] Have only one set of buttons for add module dialog
This replaces the fixDialogButtons helper that was causing a second
'hidden' set of buttons to be displayed with styles for form-dialog.

Ideally it would be nice to use react-modal, but the modules page
is sufficiently complex that its a bit difficult to do so.

closes CNVS-22870

Test Plan:
  - Go to the modules page
  - Click the 'Add Module' button
  - Using a JAWS make sure that only one set of buttons
    is read.  This should be the case in all screenreaders though
  - Close the dialog
  - The screenreader form navigation listing should not show the
    buttons as still being there.

Change-Id: I4ef0f782bb393241d1ce4bccd61ce90a0e1ba00f
Reviewed-on: https://gerrit.instructure.com/64805
Reviewed-by: Sterling Cobb <sterling@instructure.com>
Tested-by: Jenkins
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>