add grouped parameter to SubmissionApiController#index
If an assignment is marked to be graded as a group,
speedgrader will group the assignments accordingly, but
in the API, submissions were being returned individually
rather than being grouped. This adds a check for a 'grouped'
param in the SubmissionApiController#index action so that
a single assignment/submission object is returned
per student group.

fixes CNVS-23203

Test Plan:
1. Have a course with a few students enrolled
2. Create user groups and put at least two students
in one group.
3. Create an assignment set to be graded as a group
4. Have a student from one of the groups submit
to the assignment on behalf of the group.
5. With a teacher or admin token, ping the
Submissions list API with params grouped = true
Verify just a single assignment object is returned

Change-Id: I3465b469533d6cff6b3b7c679d8868ceef3dee17
Reviewed-on: https://gerrit.instructure.com/65312
Reviewed-by: Spencer Olson <solson@instructure.com>
Tested-by: Jenkins
Product-Review: Spencer Olson <solson@instructure.com>
QA-Review: Jason Carter <jcarter@instructure.com>