disable SRGB's assignment editing for past grading periods
Fixes CNVS-22398

CNVS-22398 was previously fixed with c428f832, however,
that commit created issues in Continious Integration, with
failures which only happened when the tests were run in
parallel. c428f832 was reverted in f566717b.

This commit differs from the c428f832 by avoiding
using fakeENV to manage creating an ENV for testing
purposes. The function isInPastGradingPeriodAndNotAdmin
is not as well tested as it was in c428f832, but this is
preferable to having intermittent spec failures in CI.

Test Plan:

1. As an admin, create an account-level grading period
    which ends in the past.
2. Create an assignment that is due within that grading
    period.
3. As a teacher, go to Gradebook, click on the grades for
    that assignment.
4. Observe that they are not editable.
5. Navigate to Individual View. Select the past grading
    period.
6. Select the assignment and a student in the Content
    Selection section.
7. Verify that you are unable to edit this grade.
8. Click "Submission Details."
9. Verify that you are unable to edit this grade.

Change-Id: I381c3d3c78f00eabc7a685ef3a0e8fea0f886ac2
Reviewed-on: https://gerrit.instructure.com/65417
Tested-by: Jenkins
Reviewed-by: Derek Bender <djbender@instructure.com>
QA-Review: Jason Carter <jcarter@instructure.com>
Product-Review: Christi Wruck