Start using Babel to compile files in the JSX folder
This replaces the deprecated JSX compiler formerly provided by
react-tools to use the Babel preprocessor.

This allows us to use ES6+ inside Canvas :)

Unfortunately, we had the jsx pragma hanging around all over the place
and Babel considers this an ERROR now... so I ripped them all out.

Babel also turns on strict mode for JS files, which means I needed
to fix a few things in other places to make it work such as duplicate
object keys and variables without declaration. DueDateTokenWrapper.jsx
had a problem with a 'this' being seen on the global level which Babel
detects to be nothing so it throws in an undefined. I reworked it so
that the function should still have the intended effect and 'this'
is properly scoped.

closes CNVS-24123

Test Plan
  - Make sure all JSX compiles via Guard and via js:jsx rake task
  - Make sure nothing is broken in any of the react-y areas of
    Canvas.

Change-Id: I5ab8f7f435504f8269e2b7da96813337676cf175
Reviewed-on: https://gerrit.instructure.com/65451
Reviewed-by: Benjamin Porter <bporter@instructure.com>
Tested-by: Jenkins
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>