add links to the course search tab
fixes CNVS-25165
fixes CNVS-25176

Add links to the rows, also restructure the dom so it can use flex.
Adds the analytics button if it can be added as well.

Test Plan

Given you are set up with the new Course/People search page
And you're on that page
When you naviage that to that page
And you have the analytics plugin installed
And its enabled
Then you should see a button for 'Analytics' in the top right corner

Given you are set up with the new Course/People search page
And you're on that page
When you naviage that to that page
And you have a list of courses
Then the links on those courses should work and take you to the right
place

Change-Id: I36a4b9a0033a7c441b0e9abae05559dbcdd78d61
Reviewed-on: https://gerrit.instructure.com/68989
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>
Tested-by: Jenkins
QA-Review: Clare Strong <clare@instructure.com>