remove student visibility of comments when assignment is muted
Previously, when an assignment was muted, teacher comments
were hidden from the student inconsistenly in a few pages.
The commments were hidden in the submission details page,
but not from the student grades page or the assignment
show page sidebar. Added logic to to the erb templates
so that the muted status of the assignment is being
checked first before displaying comments.

fixes CNVS-20047

Test Plan:
1. create an assignment and mute it
2. as the instructor add a submission comment in speedgrader
3. unmute the assignment
4. go back in and re-mute the assignment
5. As a student, go to the assignment show page
(/courses/x/assignments/y) and verify no comments
are visible in the sidebar, but rather, there is a
message about "You may not see all comments...."
6. Also go to the 'Grades' page (/courses/x/grades)
and verify no comments are visible for the muted assignment.
7. Also go to the submission details page (click on the name
of the assignmet in the 'Grades' page to go there) and
verify no comments are visible. Rather, there is a message
about "You may not see all comments...."
8. Unmute the assignment and make sure that the comments
show up in the pages from steps 5 - 7.
9. Click the 'Show All Details' button and make sure that
comments/rubrics for muted assignments do not show up.

Change-Id: Iba22ced513bdd6ba5aa3a1b939675c90b45fcb7c
Reviewed-on: https://gerrit.instructure.com/68210
Reviewed-by: Spencer Olson <solson@instructure.com>
Tested-by: Jenkins
QA-Review: Jason Carter <jcarter@instructure.com>
Product-Review: Christi Wruck