Fix file preview a11y issues
Makes it so that the info panel button indicates to
screenreader users that the info panel is displayed
so that they can then navigate to it.

It also hides the next/previous arrow icons from
screenreaders that way they don't get some wrong
indication that a picture is there.  I mean, there
is a picture there... but it's only for visual
presentation and doesn't need to be read because
there is other text there for screenreader users.

fixes CNVS-25696
fixes CNVS-25824

Test Plan:
  - Go to Files
  - Open a file preview
  - Enable a screenreader
  - Navigate to the "Info" button
  - Pressing it should indicate 'Info panel displayed'
  - Pressing it again should indicate 'Info panel hidden'
  - The arrows to navigate through the available files
    should indicate next and previous file as appropriate
    to the screenreader.
  - The arrows should not indicate the presence of an
    image.

Change-Id: I115172f69b117e349598d74016e96040d2e837c6
Reviewed-on: https://gerrit.instructure.com/68854
Tested-by: Jenkins
Reviewed-by: Jon Willesen <jonw@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>