fix canvas modals accessibility issues
fixes CNVS-25910

When ReactModal was upgraded, it introduced a new API and changes some
behavior with setAppElement. Now, react modal accepts a property,
appElement and needs to be set. By default, it attaches to body,
however in canvas this is really bad because it will aria-hide the
entire application whenever a modal opens. This commit makes the
default attachment #application with an option to pass in a different
element if you'd like.

Test Plan

Make sure opening modals do NOT attach aria-hidden to the <body>
element. There are many places where canvas modal is used. Here are a
few know places. If they work, then its safe to assume its working
everywhere else.

- File Rename Modal (on the files page)
- File Move Dialog (on the files page)
- File Zip/Unzip Dialog (on the files page)

Test thoses spots, ensure aria-hidden is not on the body, instead it
should be attached on the div with an id of #application.

Change-Id: I606e46524b385da85f8b9cce4f41fd7449c9dec2
Reviewed-on: https://gerrit.instructure.com/68964
Tested-by: Jenkins
Reviewed-by: Andrew Butterfield <abutterfield@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>