Update masquerade and student view alert box
This makes the alert box into a banner showing at the bottom of
the screen.  It is now magenta in color ($ic-color-action).  This
commit also makes a border show around the body of the page showing
the same color.

This will keep the dialog out of the way of the rest of the content
on the page.

closes CNVS-26723

Test Plan:
  - Masquerade as a user
  - The banner should show at the bottom in a magenta color.
  - The screen should have a magenta border around it.
  - As a teacher, Go to the student view of a course
  - Similar behavior should be shown
  - Go to a module item that has a module sequence footer.
  - The module sequence footer should not be obstructed by the
    banner.

Change-Id: I54fb6e482c75fdbc80de974f887ec6ed285131c9
Reviewed-on: https://gerrit.instructure.com/71058
Reviewed-by: Colleen Palmer <colleen@instructure.com>
Product-Review: Colleen Palmer <colleen@instructure.com>
Tested-by: Jenkins
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>