fix media_object_failure errors
Our video platform stopped transcoding to .flv a long time ago, so every
single new upload was failing this check for .flv and generating an
error report. What we really want is to ensure that the mp3/mp4
transcoding is available, depending on if it is audio or video.

This error didn't prevent the media from actually embedding and playing,
which is why it was never really noticed before. So this commit is
effectively just silencing noise.

test plan: enable media uploads and upload a video/audio file. it should
transcode and play as expected, and no media_object_failure error should
be recorded.

Change-Id: I960980ecff8a8235b0598b60fd257b22e9cd5f1f
Reviewed-on: https://gerrit.instructure.com/71142
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins
QA-Review: August Thornton <august@instructure.com>
Product-Review: Brian Palmer <brianp@instructure.com>