Fix layout for assignment submission details page
Fixes: CNVS-26813

Note: This is really old code, with all inline
styles; so to fix the bug I had to create a new
stylesheet for this view and move the affected
inline styles to classes. This feature needs a
whole bunch of additional updating; but for now,
these changes address the problem in the ticket and
overall make the page a little nicer to use in the
new UI.

Also, this page iframes Canvas inside itself, so
I created a new environment variable that we can
use to generate a class for the body; so we can
style only the Canvas that is inside the iframe.

Test plan:

- Turn on the new UI

- As a student or teacher, create or go to an
  assignment submission and click on the
  Submission Details link in the top-right box.

- For the new UI, the changes in this commit are
  as follows:
  Before: http://screencast.com/t/LoPnF8F59
  After: http://screencast.com/t/5aWTI7o9s

- For legacy UI, the changes in this commit are
  minimal:
  http://screencast.com/t/gIXRi3xA7aUP

Change-Id: I95ff8586c479b1d8ac049fb3f501b3ef8e961eba
Reviewed-on: https://gerrit.instructure.com/72075
Tested-by: Jenkins
Reviewed-by: Colleen Palmer <colleen@instructure.com>
Product-Review: Colleen Palmer <colleen@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>