fix coffeescript syntax for future versions
CoffeeScript source is currently pinned to 1.6.2 but these two
files will have syntax errors if/when coffeescript is upgraded
to >=1.7.0

To demonstrate the syntax errors, temporarily modify `Gemfile.d/development_and_test.rb`
to use `1.7.0` for `coffee-script-source`, bundle update, and
recompile assets.

These are the only 2 errors and will work with coffeescript 1.10 as well.

Test Plan
- There should be no change. I did not see if this code is currently covered by a test case.

This is a low-priority PR. I needed to track down these two
problems to use a gem which required a higher version of coffee-script.