log external tool assignment views as assignment AUAs
Our AUA system is set up to be 1-1 with requests, but in the case of
external tool assignments, we call log_asset_access twice, once for the
assignment and once for the tool, and the system is set up that the last
call wins. It has been determined that the old behavior of counting
these as assignments is a better fit.

However, this system is tied into the Live Events system, and that
system handles multiple accesses per request just fine, so we want to
continue to log these to that system.

fixes CNVS-26743

test plan:
- view assignment and non-assignment external tools as a student
- look at the results in the course user page "Access Report"
- the assignment views should be counted as assignments

Change-Id: I31687222065a99c7680012f240b799dd7485edc9
Reviewed-on: https://gerrit.instructure.com/72280
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>