fix wiki sidebar new wiki page permission check
it was relying on an ivar to exist that wasn't always set.  because of
caching, whichever environment it rendered in first for a given user
(with or with @wiki) would persist across the course.

fixes CNVS-27726

test plan:
- clear your redis cache
- visit the a page with the wiki sidebar
- it should correctly show "Link to a new page" based on whether you
  have permission or not

Change-Id: I52ea02075f8d159e0ca159eb0852a140cdfc5172
Reviewed-on: https://gerrit.instructure.com/73765
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>